tip and proper split
