+++ code/xcvm/cosmwasm/contracts/gateway/src/contract/ibc/mod.rs
@@ -1,0 +1,10 @@
+use crate::events::make_event;
+
+pub mod ics20;
+pub mod one;
+
+pub fn make_ibc_failure_event(reason: String) -> cosmwasm_std::Event {
+	make_event("receive")
+		.add_attribute("result", "failure")
+		.add_attribute("reason", reason.to_string())
+}
