+++ code/xcvm/cosmwasm/contracts/gateway/src/lib.rs
@@ -5,7 +5,5 @@ pub mod auth;
 pub mod contract;
 pub mod error;
 mod events;
-pub mod exec;
 mod prelude;
 pub mod state;
-mod topology;
